historical_data/requirements.txt
requests==2.32.3
pandas==2.2.3
boto3==1.37.18
PyYAML==6.0.1
awswrangler==3.11.0